+++ lib/krikri/parsers/oai_dc_parser.rb
@@ -4,6 +4,8 @@ module Krikri
   # metadata path as harvested from OAI-PMH.
   # @see Krikri::XmlParser
   class OaiDcParser < XmlParser
+    include Krikri::OaiParserHeaders
+
     def initialize(record, root_path = '//oai_dc:dc')
       super
     end