+++ spec/lib/krikri/harvesters/oai_harvester_spec.rb
@@ -236,7 +236,7 @@ describe Krikri::Harvesters::OAIHarvester do
       it 'saves harvest options correctly when creating an activity' do
         # Ascertain that options particular to this harvester type are
         # serialized and deserialized properly.
-        described_class.enqueue(Krikri::HarvestJob, opts = args)
+        described_class.enqueue(opts = args)
         activity = Krikri::Activity.first
         opts = JSON.parse(activity.opts, symbolize_names: true)
         expect(opts).to eq(args)